Move app events to sagas
