+++ src/main.js
@@ -1,7 +1,6 @@
 import { setupDevelopmentTools } from './main/dev';
 import { setupErrorHandling } from './main/errors';
 import { handleStartup } from './main/startup';
-import { setupAppEvents } from './main/events';
 import { setupReduxStore } from './main/reduxStore';
 
 if (require.main === module) {
@@ -9,5 +8,4 @@ if (require.main === module) {
 	setupErrorHandling();
 	handleStartup();
 	setupReduxStore();
-	setupAppEvents();
 }
